+++ framework/filters/AccessRule.php
@@ -36,6 +36,11 @@ class AccessRule extends Component
      */
     public $controllers;
     /**
+     * @var array list of module IDs that this rule applies to. The comparison is case-sensitive.
+     * If not set or empty, it means this rule applies to all modules.
+     */
+    public $modules;
+    /**
      * @var array list of roles that this rule applies to. Two special roles are recognized, and
      * they are checked via [[User::isGuest]]:
      *
@@ -105,6 +110,7 @@ class AccessRule extends Component
             && $this->matchIP($request->getUserIP())
             && $this->matchVerb($request->getMethod())
             && $this->matchController($action->controller)
+            && $this->matchModule($action->controller->module)
             && $this->matchCustom($action)
         ) {
             return $this->allow ? true : false;
@@ -128,7 +134,16 @@ class AccessRule extends Component
      */
     protected function matchController($controller)
     {
-        return empty($this->controllers) || in_array($controller->uniqueId, $this->controllers, true);
+        return empty($this->controllers) || in_array($controller->id, $this->controllers, true);
+    }
+
+    /**
+     * @param \yii\base\Module $module the module
+     * @return boolean whether the rule applies to the module
+     */
+    protected function matchModule($module)
+    {
+        return empty($this->modules) || in_array($module->uniqueId, $this->modules, true);
     }
 
     /**